worker-python/requirements.txt
Flask==2.2.2
Pillow==9.4.0
Werkzeug==2.3.8